1. Implemented the animateTransform element. 2. Added animation support to SVGOMAnimatedTransformList. 3. The transform attribute is now animatable.
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/batik/branches/anim@420104 13f79535-47bb-0310-9956-ffa450edef68

+++ sources/org/apache/batik/anim/TransformAnimation.java
@@ -32,7 +32,7 @@ public class TransformAnimation extends SimpleAnimation {
      * The transform type.  This should take one of the constants defined
      * in {@link org.w3c.dom.svg.SVGTransform}.
      */
-    protected int type;
+    protected short type;
 
     /**
      * Creates a new TransformAnimation.
@@ -48,7 +48,7 @@ public class TransformAnimation extends SimpleAnimation {
                               AnimatableValue from,
                               AnimatableValue to,
                               AnimatableValue by,
-                              int type) {
+                              short type) {
         super(timedElement, animatableElement, calcMode, keyTimes, keySplines,
               additive, cumulative, values, from, to, by);
     }
